// frontend/src/apis/models/ChildAccountResponse.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * TellMom API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ChatPlatform } from './ChatPlatform';
import {
    ChatPlatformFromJSON,
    ChatPlatformFromJSONTyped,
    ChatPlatformToJSON,
    ChatPlatformToJSONTyped,
} from './ChatPlatform';

/**
 * 
 * @export
 * @interface ChildAccountResponse
 */
export interface ChildAccountResponse {
    /**
     * 
     * @type {number}
     * @memberof ChildAccountResponse
     */
    id: number;
    /**
     * 
     * @type {ChatPlatform}
     * @memberof ChildAccountResponse
     */
    platform: ChatPlatform;
    /**
     * 
     * @type {string}
     * @memberof ChildAccountResponse
     */
    platformUserId: string;
    /**
     * 
     * @type {string}
     * @memberof ChildAccountResponse
     */
    displayName: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ChildAccountResponse
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof ChildAccountResponse
     */
    updatedAt: Date;
}



/**
 * Check if a given object implements the ChildAccountResponse interface.
 */
export function instanceOfChildAccountResponse(value: object): value is ChildAccountResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('platform' in value) || value['platform'] === undefined) return false;
    if ((!('platformUserId' in value) && !('platform_user_id' in value)) || (value['platformUserId'] === undefined && value['platform_user_id'] === undefined)) return false;
    if ((!('displayName' in value) && !('display_name' in value)) || (value['displayName'] === undefined && value['display_name'] === undefined)) return false;
    if ((!('createdAt' in value) && !('created_at' in value)) || (value['createdAt'] === undefined && value['created_at'] === undefined)) return false;
    if ((!('updatedAt' in value) && !('updated_at' in value)) || (value['updatedAt'] === undefined && value['updated_at'] === undefined)) return false;
    return true;
}

export function ChildAccountResponseFromJSON(json: any): ChildAccountResponse {
    return ChildAccountResponseFromJSONTyped(json, false);
}

export function ChildAccountResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChildAccountResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'platform': ChatPlatformFromJSON(json['platform']),
        'platformUserId': json['platform_user_id'],
        'displayName': json['display_name'],
        'createdAt': (new Date(json['created_at'])),
        'updatedAt': (new Date(json['updated_at'])),
    };
}

export function ChildAccountResponseToJSON(json: any): ChildAccountResponse {
    return ChildAccountResponseToJSONTyped(json, false);
}

export function ChildAccountResponseToJSONTyped(value?: ChildAccountResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'platform': ChatPlatformToJSON(value['platform']),
        'platform_user_id': value['platformUserId'],
        'display_name': value['displayName'],
        'created_at': value['createdAt'].toISOString(),
        'updated_at': value['updatedAt'].toISOString(),
    };
}